requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pylint==2.3.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
typed-ast==1.4.0
urllib3==1.25.5
wrapt==1.11.2
Flask==1.1.1
markdown==2.6.11
flask_restful==0.3.6
pymongo==3.10.1
dnspython==1.16.0
gunicorn
pytest==5.2.0
selenium==3.141.0
python-dotenv==0.13.0